Merge 121600 - Increase the rate at which we check
This is because now that we have several components available, whenever we successfuly download
one we would sleep for 4 hours, no matter if other components could be installed.

With this change we only sleep for 1 hour, once all components are up-to-date or have no updates we
check only after MinimumReCheckWait()

So as long there are no components to install the only thing that this change does is to
post a task to the UI four times as fast as before, e.g. every hour.

BUG=113749
TEST=see bug.

Review URL: http://codereview.chromium.org/9380031

TBR=cpu@chromium.org
Review URL: https://chromiumcodereview.appspot.com/9617037