Generalize *OnAuthRefresh and *OnCompleted methods in DocumentsService.
This CL introduces a GDataOperationInterface interface for abstracting
the interactions between DocumentsService and UrlFetchOperation<T>.
Operation classes derived from UrlFetchOperation<T> are modified to
implement the GDataOperationInterface interface.

Generic methods for handling GDataOperationInterface instances
(StartOperationToUIThread, StartOperation, RetryOperation, and
OnOperationAuthRefresh) are added to DocumentsService to replace
*OnAuthRefresh and *OnCompleted methods.

This CL also fixes an issue with the base::Value* data not being freed
in GDataOperationCallback. The solution is to explicitly transfer the
ownership of the data as scoped_ptr<base::Value> to the callback.

BUG=none
TEST=unit_tests --gtest_filter='*GData*' and manual testing in file manager
(TODO: Make DocumentsService methods mockable and add unit tests)

Review URL: http://codereview.chromium.org/9594008