Enable audio/video tag in content_shell
Some media initialization work needed to be added to the shell in order for
the ffmpeg stubs to be loaded and for the media tags to be recognized.

AudioManager also needed to be added to the shell, as the Chromium-equivalent
lives in BrowserProcessImpl, and the media code should not assume the
MediaInternals has been created.

BUG=112043,116906
TEST=content_shell builds and can play the videos in the bug above

Review URL: http://codereview.chromium.org/9316077