gdata: Add a private extension API to get gdata files.
This initiates downloading of gdata files if these aren't cached,
and returns the local file paths to the javascript layer.

This is a reland of crrev.com/125432 that was reverted as
it broke FileSystemExtensionApiTest.* tests. The tests
were broken due to my last minute change to remove "optional": "false"
from the JSON file. There was a place where it had to be "optional": true,
but was "false", which was interpreted as true. :~(
I know what was the culprit now, and I'll fix it in a separate patch.

BUG=chromium-os:27404
TEST=manually; browser_tests --gtest_filter=FileSystemExtensionApiTest.*

Review URL: https://chromiumcodereview.appspot.com/9608022