Refactor sync tests to use in-memory database
This commit makes the DBS a pure-virtual class and introduces three
concrete sub-classes: OnDiskDirectoryBackingStore,
InMemoryDirectoryBackingStore and TestDirectoryBackingStore.  This
allows us to separate test-specific functionality (ie.
OpenConnectionForTest() and some friend declarations) from the main
DirectoryBackingStore interface.

OnDiskDirectoryBackingStore is drop-in replacement to the "classic"
DirectoryBackingStore.  InMemoryDirectoryBackingStore is an easy to use
class that is meant to be injected into test classes that require a
DirectoryBackingStore.  The TestDirectoryBackingStore is a bit more
complicated, and is used only when directly testing the
DirectoryBackingStore.

TestDirectorySetterUpper and most test fixtures now use
syncable::Directory::OpenInMemory() instead of Open().  This makes the
Directory initialize itself with InMemoryDirectoryBackingStore.  Tests
that requiring an on-disk database have been modified to use
OnDiskDirectoryBackingStore.

The DirectoryBackingStore unit tests now use TestDirectoryBackingStore.
The unit tests must manage their own sql::Connection and provide it to
the TestDirectoryBackingStore.  This allows them to inject initial
database state into the test without having to create the database on
disk.

These changes have a significant impact on the time taken to run sync
unit tests on my local machine.  The average run time has gone from over
1 minute to under 20 seconds.

BUG=116328
TEST=


Review URL: http://codereview.chromium.org/9549023