bluetooth: replace interface with new API
Replace the Chrome Bluetooth interface with a completely new API. Compared
to the previous code, this gives us a much cleaner interface for the
functionality we need and completely hides that there is BlueZ and D-Bus
underneath.

BluetoothManager is replaced by BluetoothAdapter::CreateDefaultAdapter(),
the adapter code itself is responsible for tracking the default adapter
removing the need for the singleton object that persists for the Chrome
lifetime. It's perfectly acceptable to have multiple instances of
BluetoothAdapter for one device as they share the same Bluetooth*Client
instances and dbus::ObjectProxy instances.

Neither BluetoothAdapter or BluetoothDevice rely on arbitrary "unique ids"
that were really D-Bus object paths. An owner of a BluetoothAdapter instance
may simply store and use the instance; for BluetoothDevice one just uses
the address() property to identify it.

BluetoothDevice now provides some functionality that was previously
implemented in JavaScript, such as setting the name of devices before
the real name or alias is available, and determining whether or not the
device is supported. This allows us to use in multiple places in the UI.

Support for devices that are not paired, but are connected or have been
connected without being paired in the past is implemented. This may be
used for mice, and other devices where pairing is not usual.

UI Options Handler code has been updated to use the new interfaces,
support for handling USB adapter removal and insertion, control of the
adapter radio power, and populating the list of known devices has been
added.

BUG=chromium-os:27103,chromium-os:27102,chromium-os:23209,chromium-os:26265
TEST=tested on devices

Change-Id: I4d8fa830c8d48877ce4b24fee9168ee8d5386031


Committed: http://src.chromium.org/viewvc/chrome?view=rev&revision=125725

Review URL: http://codereview.chromium.org/9572025