Fix pipelining crash when send is aborted while queued.
If a request is canceled while it's in the pending_send queue, it's not removed
from that queue. If a later request calls Close(true), which evicts the
pending_send queue, it calls back into a deleted callback. Now, we remove it
from the queue.

BUG=117291
TEST=net_unittests


Review URL: http://codereview.chromium.org/9657019