PluginLoader loads plugins again after last load completes
If the list of plugins has updated before PluginLoader gets a
request to load plugins it occasionally fails to load the entire
list. This happens when PluginLoader is serving the last plugin
load request.

To make sure PluginLoader is up-to-date with the latest plugin
list all plugin load requests are served.

BUG=117561
TEST=content_unittests --gtest_filter=PluginLoader*
R=rsesek

Review URL: http://codereview.chromium.org/9665017