Ash: Fix find-in-page background not painting
This was caused by me misunderstanding what the parameter "bool restored" meant in GetHorizontalTabStripVerticalOffset().  I thought it meant to return the value for a restored window, whereas it actually meant to override the current (possibly maximized) state of the window.  I fixed my math and renamed the variable to "force_restored" to be clearer.

BUG=117487
TEST=visual


Review URL: http://codereview.chromium.org/9691018