[Chromoting] Fix left alignment of stats bar
The stats bar assumes that it is aligned to the left side. Recent changes to the UI added a 20px margin around the body area which resulted in the stats bar being offset rather than running across the entire bottom of the window.  This CL fixes that.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/9664030