Make pipelining resilient to proxy changes.
This is a speculative fix for the crashes described in bug 117649. The symptom
is a crash when we try to add pending requests to a pipeline. Analyzing the
crash, it looks like |request_map_| is empty (meaning all pending requests have
been deleted), but |http_pipelined_request_map_| is non-empty (meaning it has a
stale pointer).

The most likely cause is the same request is added to the map twice. This would
happen if SetHttpPipeliningKey was called twice, which could happen if
DoInitConnection is called twice, which could happen if the connection fails and
ReconsiderProxyAfterError returns OK, which could happen if the proxy config had
changed while the first connection was in flight.

If this doesn't fix it, I've also added some extra checking to help nail it
down.

BUG=117649
TEST=net_unittests


Review URL: http://codereview.chromium.org/9689038