Implement BrowserPluginPlaceholder.
A browser plugin is a plugin container that hosts an out-of-process RenderView (guest).

Loading up a new process, creating a new RenderView, navigating to a given URL,
and establishing a guest-to-host channel can take hundreds of milliseconds.

Furthermore, a RenderView's associated browser-side WebContents, RenderViewHost, and
SiteInstance must be created and accessed on the UI thread thread on the browser.

Thus, we must avoid blocking the host RenderView as well to avoid introducing the potential
for a deadlock.

To address the two issues above, we use a BrowserPluginPlaceholder (currently an empty
WebViewPlugin wrapper) to take place of the guest renderer until the guest renderer is
ready.

BUG=117894
TEST=manually

Review URL: http://codereview.chromium.org/9668031