Don't perform or require a successful Chrome shutdown at the end of Chrome Frame Net Tests.
These tests launch Chrome in ways that differ substantially from the norm. Hence, recording and forcing the diagnosis/fixing of failures in Chrome internal code here is an unreasonably high bar.

The middle-term objective is to actually not launch Chrome here at all. This CL supports the short term objective of getting the tests on the waterfall.


BUG=None
TEST=chrome_frame_net_tests.exe


Review URL: http://codereview.chromium.org/9677042