Gave native File Browser file handlers higher priority than 3rd party handlers.
In the File Browser if there are several actions available for a file then the most recently used one becomes the default (it is invoked on double click etc). On a fresh profile when all timestamps are equal this is a matter of chance. 
The patch gives higher priority to the actions of the File Browser (only as long as timestamps are equal).

BUG=chromium-os:27216
TEST=Start on a fresh profile, observe that "View" is shown for an image file, not "Upload to Picasa"


Review URL: http://codereview.chromium.org/9692049