mac: Remove explicit dependency from many test targets on webcore
This was added back when we were using a 32bit linker, as a manual hint
to ld. Now that webcore is split up and we use a 64bit linker, it's no longer
necessary.

It also breaks the component build, which relies on everything just depending
on libwebkit.dylib.

BUG=43791,90078
TEST=things still build
TBR=mark

Review URL: http://codereview.chromium.org/9700055