Ash: Extend tabs to top of window for maximized windows
This makes them clickable when the cursor is at the edge of the screen.

For maximized windows we suppress the resize area that extends 1 pixel inside the window bounds, as maximized windows can't be resized (at least for now).  This allows double-clicking on the top edge of the screen to restore a maximized window and suppresses misleading resize cursors.

BUG=117952
TEST=manual, maximize a window and try single and double-clicking along the top screen edge.

Review URL: https://chromiumcodereview.appspot.com/9706044