views: Rename MenuButtonDelegate::RunMenu to something more obvious.
The reason is that there are a lot of RunMenu() functions all over the code,
and actually the event for this function is that the menu button was pressed.
So we tell that to subclasses and they do whatever they want; they do not necessarily need to run a menu.
Thus this patch renames RunMenu to OnMenuButtonClicked.

BUG=117092
R=sky@chromium.org
TBR=stevenjb@chromium.org

Review URL: https://chromiumcodereview.appspot.com/9693022