Move InMemoryURLIndex Caching Operations to FILE Thread
Reading and writing of the InMemoryURLIndex (IMUI) cache is now peformed on the FILE thread and, when required, rebuilding of the index is performed on the history thread. Reading and rebuilding of the index from the cache file or history database are now performed as atomic operations rather than by restoring/rebuilding in-place, i.e. a new private data object is created and populated then swapped in for the live private data object.

BUG=83659
TEST=Unit tests updated/enhanced.

Previously reviewed: https://chromiumcodereview.appspot.com/9030031 (This change identical to 9030031.)

TBR=brettw@chromium.org
Review URL: https://chromiumcodereview.appspot.com/9699094