[Sync] Add support for associating a new Synced Bookmarks node.
Previously we would result in an unrecoverable error when the server sends
us a newly created Synced Bookmarks node. We now properly detect and associate
it, allowing us to also associate new bookmarks created under the Synced
Bookmarks folder.

In addition, we soften our conditions for unrecoverable errors. A bookmark
that cannot be added is ignored, instead of triggering an error. There are
very few of these, and are likely triggered by a crash during initial
association. 

BUG=108978
TEST=sync_integration_tests


Review URL: http://codereview.chromium.org/9570055