Add the first part of Chrome-side support for NaCl validation caching.
This code injects a validation caching implementation into sel_ldr that hashes
all of the data it is given and then queries against a stub database.  The stub
database returns whatever value is specified in the NACL_VALIDATION_RESULT env
var for all validation requests.  This code is inactive in Chrome unless
NACL_VALIDATION_CACHE=1 is set in the environment.

BUG= http://code.google.com/p/nativeclient/issues/detail?id=2515
TEST= unit_tests --gtest_filter=NaClValidationQueryTest*

Review URL: https://chromiumcodereview.appspot.com/9553009