Fix cookie settings for URLFetcher used by sync integration tests
The sync integration tests use a temporary URLFetcher to check if the
test sync server is up and running prior to running tests. There is no
need for this fetcher to store cookies on the local machine.

This patch updates the load flags for the URLFetcher to not store
cookies.

BUG=118959, 118914
TEST=sync_integration_tests

Review URL: https://chromiumcodereview.appspot.com/9748001