bluetooth: use better definitions of Paired
Recent Bluetooth specifications (such as V4.0 and HID 1.1) redefine "paired"
to be less confusing: pairing is the process of connecting two devices and
agreeing a basic link key, even if only temporary for that session. The
term "bonded" is used for a permanent link key that can be used for the
next connection.

While BlueZ still uses the older definitions, it makes sense for our
interface to use these new definitions, since it neatly divides devices
into "paired" and "not yet paired", with "bonded" a sub-category of
"paired".

BUG=none
TEST=verified on device

Change-Id: I29fc55fb9ffd0a20f8280e2b6371c3be459467b2

R=satorux@chromium.org,jhawkins@chromium.org,csilv@chromium.org,kevers@chromium.org

Review URL: https://chromiumcodereview.appspot.com/9766005