Introduce a mechanism for creating objects in JS that are lazily evaluated.
This change allows JS source to be associated with an object, which will be
evaluated and delegated to on property access to that object. This will allow,
for example, extension APIs to be setup from C++ without executing any JS until
the API is actually used.

BUG=104100
TEST=unit tests provided


Review URL: http://codereview.chromium.org/9747002