Alternative method for dealing with zero browsers
This is an alternative method for allowing chromeos to go to zero browsers which is a bit more similar to what happens on the Mac.
This method puts a lot of overhead into adding a new call to
browser_shutdown::SetTryingToQuit(false) which there already is because of Mac behavior.

Review URL: http://codereview.chromium.org/9723001