Add a GoogleSearchCounter class to count Omnibox searches.
This listens to all committed naviations and analyzes them to see if they are a omnibox Google search. If so, this logs the appropriate enum value with GoogleSearchMetrics.

We plan on expanding this in the future to count searches from other access points to help understand the distribution of searches in Chrome.

BUG=116555
TEST=Do a search in Chrome using the omnibox with instant turned off. Check about:histograms and ensure that there is an entry for GoogleSearch.AccessPoint enum value 1.


Review URL: http://codereview.chromium.org/9655004