Increase number of max sockets per group for WebSocket connections.
This change makes ClientSocketPoolManager aware of SocketPoolType so it can
pick a different limit for normal socket pools and WebSocket socket pools.

As an initial (experimental) setting, number of max sockets per group for
WebSocket connections is raised to 30, and other limits are set to the same
value as normal socket pools. This configuration should be revisited later
when WebSocket protocol stack starts to work and a good amount of usage metrics
are collected.

BUG=118268
TEST=none


Review URL: http://codereview.chromium.org/9764003