Hopefully fix crash in GDataOperationRegistry::OnOperationFinish.
The stack signature shows that a destroyed Operation is kept in OperationIDMap, which implies duplicated Operation is inserted to the map (otherwise it shouldn't happen, because we delete iff remove-from-map).

It actually can happen when authentification is failed:

UrlFetchOperationBase::OnURLFetchComplete(HTTP_UNAUTHORIZED)
=> re_authenticate_callback_.Run
===> DocumentsService::RetryOperation

This CL remedies the case.

BUG=chromium-os:28124
TEST=unit_tests --gtest_filter='*GData*'

Review URL: https://chromiumcodereview.appspot.com/9817015