Move VideoDecoder out of media pipeline.
This is the first step to move VideoDecoder out of Filter hierarchy. In this CL, VideoDecoder is removed from the pipeline and is owned by the VideoRenderBase now. All calls to Seek()/Pause()/Play() are removed. The video decoder is flushed or stopped from the VideoRenderBase instead of from the composite filter.

In a follow-up CL, VideoDecoder will not be a Filter any more and Seek()/Pause()/Play() methods will be removed from its definition and all implementations.

BUG=108340
TEST=media_unittest, content_unittest and some basic playing of html5 video.


Review URL: http://codereview.chromium.org/9700006