ash: Improve uber tray bubble lifetimes.
This makes us extend the lifetime of an already-shown
detailed view (e.g. volume or brightness) when the key that
shows it is pressed again.  It also fixes an issue where
there was a delay after one view was hidden before another
could be shown.

BUG=119624,119745
TEST=manual

Review URL: https://chromiumcodereview.appspot.com/9808072