Suppress the mousedown / mouseup immediately after a GestureFlingCancel
When using the touchpad to fling scroll, it is desirable to stop
the fling by tapping on the touchpad. It is however not desirable
that this quick tap create mousedown / mouseup events on web contents
as that can cause undesirable side effects (like link following).

BUG=119329
TEST=by hand


Review URL: http://codereview.chromium.org/9835009