SPDY - persist SPDY settings.
HttpServerProperties stores SPDY settings in SettingsMap
whose key is settings ID and value is a (flags, value)
pair (SettingsFlagsAndValue).

The format of the persisted data is same as before (restored
the old code that persisted the data). Verified it with an older
version of the Preferences file that is very large. Verified
that we have sent the right data  (sorted and unique settings)
to the server and that duplicates were removed from the file.

Changed all the code above SpdyFramer to use the above
spdy::SettingsMap and spdy::SettingsFlagsAndValue
data structures.

When chrome sends Spdy Settings to the server, it uses
the current SpdyFramer's API calls to create Settings
frame.

TODO: 
1) Delete spdy::SpdySettings and spdy::SpdySetting
   and replace them with SettingsMap, SpdySettingsIds, 
   SpdySettingsFlags and value.
2) Add version to the persisted data.


BUG=117760
R=rch, willchan
TEST=network unit tests and browser unit tests.

Review URL: http://codereview.chromium.org/9802003