Ignore untracked files caused by bug #119528
chrome/x64 and chrome/ipch were missing from .gitignore (while they shouldn't be created in the first place, they are because of bug #119528)
Adding .gitignore entries for every solution building in the wrong place is not viable in the long run (especially that it's only temporary for this bug).

When the bug is fixed, we should remove these entries.

BUG=119528
TEST=


Review URL: http://codereview.chromium.org/9857023