Add an omnibox suggest field trial.  The purpose of this is to improve omnibox's google suggest feature (a.k.a. referred to as "prediction service" in the setting page).
This change:

Creates the omnibox suggest field trial with 20 groups.

Sends suggest/search requests to google with a CGI parameter (sugexp=chrome,mod=NUM) that indicates which group the user is assigned to.

Drops a beacon in the UMA logs (for opted-in users) to identify the experiment group the user is in.

BUG=
TEST=by hand, using wireshark to capture outgoing HTTP requests

Review URL: http://codereview.chromium.org/9794002