Aura: Add non-browser windows into the list of "Alt + Tab" cycle.
This CL changes to use the child window list of the default container as the list of "Alt + Tab" cycle, instead of the browsers' list. But the list excludes the window which is not forcusable and has modal window as a child.

This CL can enable us to focus task manager window with "Alt + Tab" switcher.

BUG=118611
TEST=manual

Review URL: http://codereview.chromium.org/9719037