Allow external protocol handlers to open in Windows 8.
This was broken due to changes to the Windows 8 APIs around default
handlers. Chrome checks if it is the default handler for the protocol
before launching any external program. Prior to this change the check
on Windows 8 would always return success. This change makes it return
UNKOWN instead.

BUG=116452
TEST=Tested external handlers work on Win 7 and Win 8. Checked register
     protocol handler still works on Win 8.


Review URL: http://codereview.chromium.org/9863026