BitmapPlatformDevice multiply-inherits from PlatformDevice and SkDevice, in that order.  However, a newly-created BitmapPlatformDevice is often upcast to an SkDevice, and owning pointers are stored to it as an SkDevice.  All the relevant destructors are virtual, so it can safely be deleted polymorphically. However, valgrind doesn't know that, and at program exit some BitmapPlatformDevices are still reachable only as SkDevices, pointing to the middle of the BitmapPlatformDevice.
The real solution would be to get rid of the MI, but for now just switching
the base classes around causes valgrind to shut up, since we're now owning
the object through the same pointer that we got from new.  This moves the 
memory in valgrind from "possibly lost" to "still reachable".

BUG=105744
TEST=see valgrind output, check that errors go away
Review URL: https://chromiumcodereview.appspot.com/9900001