Properly crash when triggering an ASSERT() in debug DRT.
We would just print the stack trace and keep running.  In webkit_unit_tests,
we would print the stack trace and crash.  Reverse the logic so in
unit_tests, we keep going and in DRT, we crash.

When rolling this into WebKit, we'll need to coordinate with the gardener
since tests will start crashing on the debug bot.

BUG=None

Review URL: https://chromiumcodereview.appspot.com/9863056