Allocate Timer object on the heap in ScreenRecorder.
base::Timer class now needs to be destroyed on the thread on which it is
started, so its lifetime needs to be separated from lifetime of
ScreenRecorder. Also removed |is_recording_| flag from that class as now
we can use capture_timer_ state instead.

BUG=120537

Review URL: https://chromiumcodereview.appspot.com/9921032