Ash: Fix maximized windows not restoring
When Chrome OS updates from R18 to R19 it will store sessions with maximized windows and restored bounds filling the screen.  The previous patch to deal with that state was interfering with recent changes around auto-hiding the launcher.  I found a better way to deal with the upgrade issue which also fixes this problem.

The problem with window restore was introduced with crrev.com/129410 but we need to retain that patch too.

BUG=120821
TEST=added unit test to BaseLayoutManagerTest
R=sky@chromium.org

Review URL: https://chromiumcodereview.appspot.com/9910024