Add linux_futura and make linux_aura more relevant today.
The idea behind this is to split the currently unclear role of the linux_aura
try bot.   Job 1: are we ready to switch to use_aura=1.   Job 2: Will this
regress the currently working stuff for the aura config.

Job 1 now becomes the mission of linux_futura.  Failures are to be expected
  for a long time, and hopefully the name conveys that.

Job 2 now becomes the duty of linux_aura.  This is the same as the predecessor
  of the same name had got to.

Doing this adds to the demand for trybots, adds more maintenance as the
aura work progresses but allows the GYP work to be minimized and deferred.

BUG=116170
TEST=local start.

Review URL: http://codereview.chromium.org/9702053