Report failures when opening an extension port to the caller.
Extension ports are now disconnected if you try to open a port to something
that has no message channel listeners. This can happen e.g. if you try to open
a port to a tab that has no content scripts injected.

Also removed usage of the generic MessageInvoke by the ExtensionMessageService. Event.dispatchJSON is the only remaining user of that.

BUG=32899,120531
TEST=no

Review URL: https://chromiumcodereview.appspot.com/9963041