Add an apiFeatures property to the client plugin.
To date features have been added with an associated plugin API bump, leading to conflicts when multiple features go in within the same API bump, and one of them is subsequently back-ported to an earlier release.

apiFeatures allows features of the client plugin to be treated independently.

TEST=Manual testing of client default scaling setting.


Review URL: http://codereview.chromium.org/9969073