Handling click on selecting checkbox in File Manager in a better way.
Currently instead of handling 'click' event we modify 'up'/'down' event on the checkbox and let selection controller to handle it. It's tricky and has side effects. The most annoying is the following. Click on the small area around the checkbox selects the item and immediattely (but visibly) unselects it.

BUG=chromium-os:28722
TEST=Manual test.

Review URL: https://chromiumcodereview.appspot.com/9963101