Clear the selection text when performing empty selections.
There is a variable in RenderViewImpl named selection_text_ that according to the description holds the text selection corresponding to the last time that didChangeSelection was called. However, it is not cleaned up when this method reports an empty selection. This can prevent detecting some selection changes and sending their update messages to the browser side.

BUG=none
TEST=none

Review URL: https://chromiumcodereview.appspot.com/9963105