Move the URL string from TemplateURLRef onto the owning TemplateURL.  This will make it easier to move the data members of TemplateURL into a new class later.
This changes the accessors for TemplateURL's TemplateURLRefs.  There are now separate accessors for the URLs as strings and as TemplateURLRefs, and the latter have changed to returning a const ref, meaning they no longer return NULL when the corresponding URL string is empty.  This makes a number of callers clearer.

BUG=none
TEST=none
Review URL: https://chromiumcodereview.appspot.com/9968016