Change SingletonHwnd to be based on WindowImpl.
This removes the need for some Windows plumbing in
SingletonHwnd and fixes an issue where a WndProc
message would be getting sent after the Singleton
destruction, which was being triggered in ui_tests.

WindowImpl is robust against the above issue because
it stores a pointer to itself in the window user data
and clears it when it is destroyed.

BUG=105550
TEST=Running ui_tests with use_canvas_skia=1 on Windows shouldn't crash.
Review URL: https://chromiumcodereview.appspot.com/9965082