Merge 130198 - [Sync] Automatically clear passphrase required reason for first time sync
In SetEncryptionPassphrase, if the passphrase required reason is
REASON_ENCRYPTION (and we're not attempting to do an invalid passphrase
transition) we know the passphrase will be accepted. As a result we can
automatically clear the passphrase required reason before the
OnPassphraseAccepted call gets sent.

BUG=121190
TEST=manual first time sync


Review URL: http://codereview.chromium.org/9968020

TBR=zea@chromium.org
Review URL: https://chromiumcodereview.appspot.com/9969127