Enforce the use of a root window
So for metro we need to have a specific root window, so we
need to enforce it. We know we are in metro because of the metro_driver
dll being loaded (it was loaded by chrome.exe)

I also removed your last change. I assume that you gathered enough data
by now.

BUG=118641
TES=chrome runs
Review URL: https://chromiumcodereview.appspot.com/9965094