Merge 130541 - [Sync] Fix reassociation when tab nodes have been only partially removed.
We would reassociate under the assumption that the number of free tab nodes
matched the tags of the tab nodes (i.e. all tab nodes have a tag based on an
id < the number of free tab nodes). When other clients delete stale sessions,
they're not guaranteed to have all the nodes, and so may delete only a subset
of the tab nodes. On the next reassociation, the previously-stale client will
not properly track the pre-existing tab nodes. Specifically, the tab nodes
it adds to its free pool may have tags based on ids larger than the free pool.

To simplify this, we just delete all pre-existing tab nodes at association time.

In addition, to lay the groundwork for future support of reassociation, we now
also have a field in each tab node that provides the actual tab id used to create the
client tag for that node. We can then eventually use that id to do tag based
lookups instead of relying on size of our tab pool.

BUG=121783
TEST=unit_tests

Review URL: https://chromiumcodereview.appspot.com/9968114

TBR=zea@chromium.org
Review URL: https://chromiumcodereview.appspot.com/9960033