Allow arbitrary types in IDL schemas
This lets us specify a type of 'object' or 'any' in IDL schemas, and optionally add
an 'instanceOf' annotation for object types that will get enforced in javascript in the same way
that the JSON schema's 'isInstanceOf' property is.

Also add some basic unit tests for IDL schemas in addition to unit tests of these new features.

BUG=119359
TEST=Included unit tests should compile and pass

Review URL: https://chromiumcodereview.appspot.com/9979014