Use the RenderText's computed text height instead of Font::GetHeight() in canvas_skia.cc.
This fixes the problem of incorrect alignment in the case of font fallback choosing a font with a different height.

Also, use the computed RenderText height to center the text in RenderText::GetOriginForDrawing(), so that the
two calculations (canvas_skia.cc and render_text.cc) use the same consistent value.

BUG=105550
TEST=Inspect views_examples text style example.
Review URL: https://chromiumcodereview.appspot.com/9923013