Merge 129977 - Call VideoDecoder::set_host() now that VD is no longer part of CompositeFilter.
r128289 moved VideoDecoder out of the CompositeFilter, but that caused it to
silently lose the Filter::set_host() call it was getting from
CompositeFilter::AddFilter() before.  This CL restores that call in the
renderer's Initialize() method.  This makes decode errors correctly report to
the Pipeline again, instead of crashing the renderer on a SEGV.

BUG=121014
TEST=playback of misdecoded file no longer crashes renderer, just greys out controls.

Review URL: https://chromiumcodereview.appspot.com/9957001

TBR=fischman@chromium.org
Review URL: https://chromiumcodereview.appspot.com/10006025