Do all Presents on present thread.
If we queue up presents from the UI thread on the present thread, that will avoid the present thread starving the UI thread due to the AcceleratedPresenter lock. The UI thread will synchronously wait for the present to finish, so the WM_PAINT callback will end at the correct time.

Also remove the spin after the present, as it's probably unnecessary.


BUG=122087
TEST=


Review URL: http://codereview.chromium.org/9949043