Enable setting the CommandBufferProxy of WebGraphicsContext3DCommandBufferImpl
By default WebGraphicsContext3DCommandBufferImpl creates a new CommandBufferProxy when it's initialized. This change introduces a new initializer that takes in CommandBufferProxy, instead.

This is used by the browser plugin to install a PpapiCommandBufferProxy instead of a CommandBufferProxyImpl that talks directly to the GPU process. 

BUG=121753
TEST=manually

Review URL: http://codereview.chromium.org/9968093