Use FILES.cfg for Windows symbols on chromium builds.
This allows us to exclude syzygy symbols which are not needed for
non-official builds, and which are just bloating our build archives and
slowing cycle times.

Note, this is only part of the change. A related change for the buildbot
scripts is in another CL.

TEST=Windows symbols archives unchanged for both Chrome and Chromium.

Review URL: https://chromiumcodereview.appspot.com/10007042