Reorder nacl_process_host functions.
No code change. This is in preparation for more significant refactoring.

This just makes the functions in the .cc file match the header, and moves the static functions to one place at the top in one anonymous namespace. This also renames the confusingly-named launch with gdb function which doesn't actually launch anything.

BUG=
TEST=

Review URL: http://codereview.chromium.org/10008055