Remove FFmpegDemuxer::first_seek_hack_ and related unnecessary code.
The warning in WebMediaPlayerImpl was added in r23598 but has long since been true.

FFmpegDemuxer::first_seek_hack_ is due to Pipeline issuing a seek to the start timestamp as the final stage of initialization, however the demuxer doesn't require prerolling so we can safely skip it and remove the hack entirely.

TEST=media_unittests, http/tests/media/, any HTTP 200 chunked response

Review URL: https://chromiumcodereview.appspot.com/10024033