This CL removes the unused trace.proto definition and build target, and makes all protobuf fields "optional".
Fields marked required cause protobuf messages to be silently ignored if those fields are not present.  By moving the fields to be optional we have more flexibility to remove fields in future.

All of the previously-required fields were in client-to-host messages; clients will need to continue to provide those fields at least until all hosts are updated not to require them - this is primarily relevant for the KeyEvent "keycode" field.

Review URL: http://codereview.chromium.org/10024003