Create a local bitmap to capture CoreGraphics rendering when the canvas is not backed by a bitmap.
During normal desktop drawing, CoreGraphics calls that
image theme elements like checkboxes and buttons draw
into the device bitmap owned by Skia's canvas. When a
PDF is created as part of a print preview, no bitmap is
allocated.

In the latter case, associate a local array of pixels
with the CGContext to capture the theme rendering, then
draw it with Skia once the theme rendering is complete.

This fix over-allocates memory. It creates a pixel map
as large as the device. The bounds of the CG drawing is
not visible to Skia, and the clip may not be set. The
PDF may be larger than need be for the same reason.

In a subsequent change, the theme element's bounds
can reduce the size of the bitmap allocation. This could
be done by setting the clip or by passing the rectangle
to the graphics context. This would require a sizeable
edit to RenderThemeMac.mm. 

Also, change the unit test to test both paths, when the
device is backed by a bitmap, and when it is not. This
requires the unit test to call the destructor and reset
the clip before comparing the results.

BUG=121752
TEST=http://stmarksparish.com/printtest.aspx, print to PDF
TEST=SkiaUtilsMacTest.BitLocker_*
Review URL: https://chromiumcodereview.appspot.com/10031005