Add a test for the "ephemeral_users_enabled" device policy
This CL adds a test for the "ephemeral users enabled" policy. The
policy prevents persistent information from accumulating by mounting
cryptohome vaults via tmpfs and not writing user preferences to the
local store. Information that previously accumulated on the system is
cleared when the policy is turned on and the first user logs in.

BUG=chromium-os:27702
TEST=New PyAuto: ChromeosEphemeralUsersEnabled.* (ChromeOS)
TEST=Affected PyAutos: prefs.PrefsTest.testHaveLocalStatePrefs,
TEST=    policy.* (ChromeOS, Linux); prefs.* (ChromeOS)

Review URL: http://codereview.chromium.org/9960074