aura/texture transport: Keep ref to texture infos to avoid races with the UI
In some cases, the UI may destroy the textures the renderer is rendering into
before the renderer has a chance to finish.
This patch makes it so that we avoid generating a lost context event in the
page that is closing, because we can't recover correctly in the other pages in
the same renderer process.

This also adds logging so that we can audit all cases of lost context.

BUG=122868
TEST=run aura chrome, open and close tabs (from the same process), observe no
flashing and no mention of a lost context.


Review URL: http://codereview.chromium.org/10103021