Convert the logging ui_test to a browser_test.
I spent a lot of time trying to find a way to make the assertion and crash dump locator tests non-flaky and shardable. But it's not really possible, because the log file is per binary, and the dumps all get put into the same directory. That means that running with sharding will give erronous results. Also even if we weren't sharding, there's still race conditions looking at the disk for a crash dump. I kept a test which looks for a notification that we crashed (instead of looking on disk). If this breaks in the field, I believe we'll notice very quickly from the canary builds (do we also have breakpad tests for this?). For the assert/check tests, I think if something breaks there developers will very quickly find it.

BUG=121574
Review URL: https://chromiumcodereview.appspot.com/10082001