ash: Go back/forward in response to buttons 8 and 9.
This is needed on Chrome OS to support both mice with
back/forward buttons and to support new events that I'll be
sending from CMT (changes in progress).

Tested that it can be triggered by the touchpad with the
"simple scrolling direction" setting honored, and that
back/forward mouse buttons work with simple scrolling ignored
in that case.

BUG=chromium-os:27349
TEST=manual: see above


Review URL: http://codereview.chromium.org/9899002