Make schema_bundle_generator escape double-quotes properly.
Currently the schema_bundle_generator will replace any double-quote character
with \", which only works if the quote is not already escaped in which case it
will escape the escaping backslash rendering the double-quote unescaped. This
patch makes sure that already-escaped double-quotes are not escaped again.

BUG=none
TEST=none


Review URL: http://codereview.chromium.org/10108029