Add nextSlaveAndBuilder to builder config
Add a function nextSlaveAndBuilder() that takes in a list of available
slaves and a list of pending buildrequests and returns the next
buildrequest to run and the slave to run it on.

This allows more advanced job scheduling, such as allowing a job to request
which slave(s) to run on (very useful for admin work), for a job to specify
a class of machines to run on, and a way to specify job priority.

An example usage is in https://chromiumcodereview.appspot.com/10031041.

BUG=chromium-os:29191
TEST=Same as Issue 10031041
Review URL: https://chromiumcodereview.appspot.com/10032026