NaCl: Remove code that attempts to kill the NaCl loader process
In a later change, I will make KillChildProcess() not work on Windows,
because the sandboxed renderer process will not get passed the process
handle for the NaCl loader proces, so the renderer cannot kill the
NaCl loader process directly.

So I'm removing KillChildProcess() for consistency between OSes and to
avoid confusion.

Also, I suspect KillChildProcess() does not work in other sandboxes
anyway:

 * In the Linux SUID sandbox, the renderers run in a different PID
   namespace, but the NaCl plugin is using a PID from the main PID
   namespace, which is wrong.

 * In the Linux seccomp sandbox (debug builds only), kill() is
   disabled.

 * I think kill() is disabled in the Mac sandbox too.

Furthermore, on Unix, if the NaCl loader process dies and its PID is
recycled, we could accidentally kill the wrong process.  (Normally the
PID cannot be recycled until the zombie process is reaped with
waitpid(), but in Chromium the NaCl loader process is a child of the
browser, not the renderer.  The NaCl process gets reaped automatically
and the NaCl plugin's call to waitpid() is useless.)

Calling KillChildProcess() is unnecessary because the NaCl loader
process is supposed to exit voluntarily when it receives EOF on its
connection to the renderer process.

Removing this call also removes the risk that we do kill(0, SIGKILL)
(if using a dummy PID of 0 later on), which would kill all processes
in the progress group and might bring down the browser.

This will let me remove KillChildProcess() from the NaCl side.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2719
TEST=nacl_integration etc.

Review URL: https://chromiumcodereview.appspot.com/10122002