Attempt to fix use after free reported by valgrind
When on_io_completed_ is signaled on file thread while
GDataFileSystem is waiting for the event in the destructor,
GDataFileSystem could free num_pending_tasks_lock_ before
the lock is released on the file thread.

Also we can avoid locking num_pending_tasks_lock_ in the
destructor if we set initial state of on_io_completed_ to be
signaled.

BUG=122717
TEST=trybots

Review URL: https://chromiumcodereview.appspot.com/10001015