Rearrange and rename about:sync details
Some of the renames are meant to clarify exactly what status the
variable is reporting.  For example, "Initial Sync Ended" has been
renamed to "Initial Download Complete".

Many of the renames and rearraning are based around an attempt to
accurately describe the lifetime and staleness of reported information.
The previous arrangement did not make it very clear which counters were
reset every cycle and which were reset only on browser restart.  The new
arrangement and descriptions should clarify this.

This change also includes fairly significant modifications to the way
that the C++ code that fetches this data communicates with the HTML side
that presents it.  Previously, the HTML code was responsible for
providing what little structure existed in the presentation of data.
This change adds the concept of "sections" of key value pairs which can
be specified by the C++ code.  This allows the sync_ui_util.cc to create
groups of key-value pairs and to name these groups.

This new functionality has been put to use in organizing the recently
renamed variables.  Some existing special case key-value pairs, like
the status summary and sync URL have been ported to this new framework
as well.

Finally, this Cl also includes changes to the way the information flows
along the page by taking advantage of the -webkit-columns property.

BUG=98346
TEST=

Review URL: http://codereview.chromium.org/9815023