Notify WebWidget when its compositor surface is ready to use
Before the browser UI thread has a chance to initialize the native window system
surface binding (which is different on each platform) any attempt to initialize
the compositor will fail in a way that appears fatal to WebKit. To avoid hitting
this, we set a bit when we receive an IPC sent by the browser UI thread after
it gets a chance to set up the native windowing stuff - RenderWidget::CompleteInit().
This bit delays RenderWidget-driven frame scheduling, but when WebKit is scheduling
frames (for example with the threaded compositor active) we also have to let
WebKit know so it can start rendering.

BUG=123356
TEST=See bug
TBR=darin


Review URL: http://codereview.chromium.org/10095021