Log when a browser plugin hangs
Add an API to webplugin_delegate_impl to get the plugin version from the window, similar to the way to get the plugin name. Update hung_plugin_action to log when a browser plugin hangs and log the version of the plugin is Google Talk Plugin. 

BUG=
TEST=try bot and manual test on Google Talk Plugin hangs


Review URL: http://codereview.chromium.org/9958104