tweaks to LocalStrings
makes some changes to LocalStrings and the templateData process. See the bug for more details. Updates the Extensions page to use the new process. Adds a test for the extensions page (I verified that if Extensions tries to access a string that doesn't exist, this test fails).

BUG=122753
TEST=included


Review URL: http://codereview.chromium.org/10054001