Convert the PPB_Flash interface to use the thunk system.
This was a bit tricky since some of the functions had no instance context, so I added global functions on the PpapiGlobals interface for them. It would be nice to add a PP_Instance argument and fix this in the future.

I did a new style of doing the thunking. The "function APIs" haven't really worked out since it's almost always easier to add the functions directly on the instance one. Since this is a larger and more separable chunk, I just added a getter on the instance API for the flash API and thunk through that. I'd like to convert the remaining ~3 function APIs to either call directly on the instance or use this method, but that's not addressed by this patch.

I moved the flash command line switch to plugin_switches so this could be hooked up properly. It allowed me to delete the delegate API for this.

I combined the flash fullscreen functions into the new Flash API which removed a bit of code.

Review URL: https://chromiumcodereview.appspot.com/10091003